REFACTOR(Web proj): structure changes
