+++ OnlineMinion.Web/Components/AccountSpecsPageEditor.razor
@@ -1,4 +1,4 @@
-@using OnlineMinion.Web.Validation
+@using OnlineMinion.Web.Components.Validation
 @{
   var name = nameof(Model.Name);
   var nameLower = name.ToLowerInvariant();
