+++ OnlineMinion.Web/Components/AccountSpecsPageEditor.razor.cs
@@ -2,8 +2,8 @@ using ErrorOr;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Forms;
 using OnlineMinion.Contracts.AppMessaging;
-using OnlineMinion.Web.Infrastructure;
-using OnlineMinion.Web.Validation;
+using OnlineMinion.Web.Components.Base;
+using OnlineMinion.Web.Components.Validation;
 
 namespace OnlineMinion.Web.Components;
