+++ OnlineMinion.Web/Components/Validation/ServerSideValidator.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Forms;
 
-namespace OnlineMinion.Web.Validation;
+namespace OnlineMinion.Web.Components.Validation;
 
 public class ServerSideValidator : ComponentBase
 {
