+++ OnlineMinion.Web/_Imports.razor
@@ -10,4 +10,4 @@
 @using OnlineMinion.Web.Shared
 @using OnlineMinion.Web.Components
 
-@inherits OnlineMinion.Web.Infrastructure.ComponentWithCancellationToken
+@inherits OnlineMinion.Web.Components.Base.ComponentWithCancellationToken
